# webui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ManManKan Local API",
    "version": "0.0.6.9.35"
  },
  "paths": {
...
    "/api/v1/boards/trends": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Board Trends",
        "description": "\u8fd4\u56de\u884c\u4e1a / \u9898\u6750\u6307\u6570\u7684\u8fde\u7eed\u8d8b\u52bf\u5feb\u7167\u3002",
        "operationId": "board_trends_api_v1_boards_trends_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/BoardKind",
              "default": "industry"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/BoardTrendMode",
              "default": "close"
            }
          },
          {
            "name": "up",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 30,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Up"
            }
          },
          {
            "name": "down",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 30,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Down"
            }
          },
          {
            "name": "min_streak",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 30,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Streak"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/BoardTrendSort",
              "default": "streak"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3,
              "minimum": 1,
              "default": 1,
              "title": "Level"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardTrendSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/boards/{kind}/{value}/pulse": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Board Pulse",
        "description": "\u8fd4\u56de\u677f\u5757\u6210\u5206\u80a1\u5728\u6700\u65b0\u5b8c\u6574\u4ea4\u6613\u65e5\u7684\u6da8\u8dcc\u7ed3\u6784\u3002",
        "operationId": "board_pulse_api_v1_boards__kind___value__pulse_get",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/BoardKind"
            }
          },
          {
            "name": "value",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Value"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3,
              "minimum": 1,
              "default": 1,
              "title": "Level"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10,
              "minimum": 1,
              "default": 5,
              "title": "Limit"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardPulseSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/board-reviews": {
      "post": {
        "tags": [
          "v1"
        ],
        "summary": "Create Board Review",
        "description": "\u4fdd\u5b58\u884c\u4e1a\u4e0e\u9898\u6750\u6700\u65b0\u8d8b\u52bf\u4e8b\u5b9e\uff0c\u5e76\u4e0e\u4e0a\u4e00\u4efd\u540c\u53e3\u5f84\u8bb0\u5f55\u6bd4\u8f83\u3002",
        "operationId": "create_board_review_api_v1_board_reviews_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardDailyReviewRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDailyReview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Board Reviews",
        "operationId": "board_reviews_api_v1_board_reviews_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BoardDailyReviewSummary"
                  },
                  "title": "Response Board Reviews Api V1 Board Reviews Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/board-reviews/{review_id}": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Board Review",
        "operationId": "board_review_api_v1_board_reviews__review_id__get",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Review Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDailyReview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stocks/{symbol}": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Stock Research",
        "operationId": "stock_research_api_v1_stocks__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockResearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stocks/{symbol}/history": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Stock History",
        "operationId": "stock_history_api_v1_stocks__symbol__history_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 360,
              "minimum": 2,
              "default": 60,
              "title": "Period"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Stock History Api V1 Stocks  Symbol  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/portfolio/cash": {
      "put": {
        "tags": [
          "v1"
        ],
        "summary": "Update Portfolio Cash",
        "operationId": "update_portfolio_cash_api_v1_portfolio_cash_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolio/positions": {
      "post": {
        "tags": [
          "v1"
        ],
        "summary": "Create Portfolio Position",
        "operationId": "create_portfolio_position_api_v1_portfolio_positions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolio/positions/{symbol}": {
      "delete": {
        "tags": [
          "v1"
        ],
        "summary": "Delete Portfolio Position",
        "operationId": "delete_portfolio_position_api_v1_portfolio_positions__symbol__delete",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BoardKind": {
        "type": "string",
        "enum": [
          "industry",
          "theme"
        ],
        "title": "BoardKind"
      },
...
      "BoardTrendMode": {
        "type": "string",
        "enum": [
          "close",
          "candle"
        ],
        "title": "BoardTrendMode"
      },
...
      "BoardTrendSnapshot": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "title": "Schema Version",
            "default": 1
          },
          "query": {
            "$ref": "#/components/schemas/BoardTrendQuery"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "data_cutoff": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Cutoff"
          },
          "partial": {
            "type": "boolean",
            "title": "Partial",
            "default": false
          },
          "coverage": {
            "$ref": "#/components/schemas/BoardTrendCoverage"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/BoardTrendRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "failures": {
            "items": {
              "$ref": "#/components/schemas/BoardTrendFailure"
            },
            "type": "array",
            "title": "Failures"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query",
          "source",
          "coverage"
        ],
        "title": "BoardTrendSnapshot"
      },
      "BoardTrendSort": {
        "type": "string",
        "enum": [
          "streak",
          "latest",
          "moneyflow"
        ],
        "title": "BoardTrendSort"
      },
...
    }
  }
}